fix LP64 portability bugs across 24 files — zero HIGH findings remain
Replace raw byte-offset pointer arithmetic (*(Type*)((char*)ptr + N))
with proper member access across 20 source files. These hardcoded PPC
32-bit struct offsets read wrong fields on x86_64 where vtable pointers
grow from 4 to 8 bytes, shifting all member offsets.

Key fixes:
- SaveLoadManager: this+0x50 → mSigninMask (renamed from unk50)
- HamDirector: mMoveMerger+0x40 → Mergers() getter
- VelocityBuffer: int unk0[2000] → const RndMesh* mMeshPtrs[2000]
- Text.cpp: 14 wchar_t/L"..." fixes for 2-vs-4 byte element size
- Ribbon/Spline/Flanger: (int)ptr → (intptr_t) address truncation
- CursorPanel: memcpy at 0x74 → TexXfm()/SetTexXfm()
- BinkIntegration: #ifdef HX_NATIVE with proper BINKIO struct access
- ContentMgr_Xbox: this+0x84/88/8c → &mEnumHandles[4/5/6]

Scanner enhancements (lp64_scanner.py):
- raw_byte_member_access: *(T*)((char*)p + N) at any offset
- ptr_int_arithmetic: (int)ptr inside +/- expressions
- variable_byte_offset_deref: *(T*)((char*)p + var)

+++ src/system/hamobj/MoveDir.cpp
@@ -120,7 +120,7 @@ namespace {
 
     void DrawBeatLine(float, float, float, const Hmx::Color &);
 
-    float DrawPlayClip(float y, SkeletonClip *portClip, int clipAsInt) {
+    float DrawPlayClip(float y, SkeletonClip *portClip, intptr_t clipAsInt) {
         SkeletonClip *clip = (SkeletonClip *)clipAsInt;
         MILO_ASSERT(clip, 0x762);
         String str(clip->Name());
@@ -1746,7 +1746,7 @@ float MoveDir::UpdateOverlay(RndOverlay *overlay, float y) {
     // Draw play clip overlay if present
     SkeletonClip *playClip = mPlayClip;
     if (playClip) {
-        y = DrawPlayClip(y, playClip, (int)(SkeletonClip *)mPlayClip);
+        y = DrawPlayClip(y, playClip, (intptr_t)(SkeletonClip *)mPlayClip);
     }
 
     // Draw filter version name
